refactor: AnalyticsChart の重複ロジックを統合
- DailyChart, BySiteChart, CumulativeChart を個別コンポーネントに分割
- chartUtils.ts に共通ロジック集約（formatMinutes, formatDate, getTimeAxisConfig）
- AnalyticsChart.tsx はチャートタイプ切り替えのみ管理（395行 → 216行）
- Y軸フォーマット処理を共通化
